+++ skyplane/cli/cli_init.py
@@ -4,15 +4,20 @@ import shutil
 import subprocess
 import traceback
 from pathlib import Path
+from typing import List
 
 import questionary
 import typer
+from rich import print as rprint
 from rich.progress import Progress, SpinnerColumn, TextColumn
-from typing import List
 
+import skyplane.api.usage
 from skyplane import compute
+from skyplane.cli.impl.common import print_header
+from skyplane.api.usage import UsageClient, UsageStatsStatus
 from skyplane.config import SkyplaneConfig
-from skyplane.config_paths import aws_config_path, gcp_config_path
+from skyplane.config_paths import aws_config_path, gcp_config_path, config_path
+from skyplane.utils import logger
 
 
 def load_aws_config(config: SkyplaneConfig, non_interactive: bool = False) -> SkyplaneConfig:
@@ -345,3 +350,66 @@ def load_gcp_config(config: SkyplaneConfig, force_init: bool = False, non_intera
                 return disable_gcp_support()
     else:
         return disable_gcp_support()
+
+
+def init(
+    non_interactive: bool = typer.Option(False, "--non-interactive", "-y", help="Run non-interactively"),
+    reinit_azure: bool = False,
+    reinit_gcp: bool = False,
+    disable_config_aws: bool = False,
+    disable_config_azure: bool = False,
+    disable_config_gcp: bool = False,
+):
+    """
+    It loads the configuration file, and if it doesn't exist, it creates a default one. Then it creates
+    AWS, Azure, and GCP region list configurations.
+
+    :param reinit_azure: If true, will reinitialize the Azure region list and credentials
+    :type reinit_azure: bool
+    :param reinit_gcp: If true, will reinitialize the GCP region list and credentials
+    :type reinit_gcp: bool
+    :param disable_config_aws: If true, will disable AWS configuration (may still be enabled if environment variables are set)
+    :type disable_config_aws: bool
+    :param disable_config_azure: If true, will disable Azure configuration (may still be enabled if environment variables are set)
+    :type disable_config_azure: bool
+    :param disable_config_gcp: If true, will disable GCP configuration (may still be enabled if environment variables are set)
+    :type disable_config_gcp: bool
+    """
+    print_header()
+
+    if non_interactive:
+        logger.warning("Non-interactive mode enabled. Automatically confirming interactive questions.")
+
+    if config_path.exists():
+        logger.debug(f"Found existing configuration file at {config_path}, loading")
+        cloud_config = SkyplaneConfig.load_config(config_path)
+    else:
+        cloud_config = SkyplaneConfig.default_config()
+
+    # load AWS config
+    typer.secho("\n(1) Configuring AWS:", fg="yellow", bold=True)
+    if not disable_config_aws:
+        cloud_config = load_aws_config(cloud_config, non_interactive=non_interactive)
+
+    # load Azure config
+    typer.secho("\n(2) Configuring Azure:", fg="yellow", bold=True)
+    if not disable_config_azure:
+        cloud_config = load_azure_config(cloud_config, force_init=reinit_azure, non_interactive=non_interactive)
+
+    # load GCP config
+    typer.secho("\n(3) Configuring GCP:", fg="yellow", bold=True)
+    if not disable_config_gcp:
+        cloud_config = load_gcp_config(cloud_config, force_init=reinit_gcp, non_interactive=non_interactive)
+
+    cloud_config.to_config_file(config_path)
+    typer.secho(f"\nConfig file saved to {config_path}", fg="green")
+
+    # Set metrics collection by default
+    usage_stats_var = UsageClient.usage_stats_status()
+    if usage_stats_var is UsageStatsStatus.DISABLED_EXPLICITLY:
+        rprint(skyplane.api.usage.USAGE_STATS_DISABLED_MESSAGE)
+    elif usage_stats_var in [UsageStatsStatus.ENABLED_BY_DEFAULT, UsageStatsStatus.ENABLED_EXPLICITLY]:
+        rprint(skyplane.api.usage.USAGE_STATS_ENABLED_MESSAGE)
+    else:
+        raise Exception("Prompt message unknown.")
+    return 0
