+++ skyplane/cli/impl/progress_bar.py
@@ -3,7 +3,7 @@ from typing import List
 from rich.progress import Progress, SpinnerColumn, TextColumn, BarColumn, DownloadColumn, TransferSpeedColumn, TimeRemainingColumn
 from skyplane import exceptions
 from skyplane.chunk import Chunk
-from skyplane.cli.common import console, print_stats_completed
+from skyplane.cli.impl.common import console, print_stats_completed
 from skyplane.utils.definitions import format_bytes
